+++ backend/pong/views.py
@@ -6,7 +6,7 @@ from rest_framework.response import Response
 from rest_framework.request import Request
 from rest_framework.views import APIView
 
-from users.serializers import SuccessResponseSerializer
+from users.serializers.UserProfile import SuccessResponseSerializer
