--- backend/users/forms.py
@@ -1,14 +1,0 @@
-from django import forms
-from django.contrib.auth.forms import UserCreationForm
-from .models import UserProfile
-
-
-class RegisterForm(UserCreationForm):
-    class Meta:
-        model = UserProfile
-        fields = ['username', 'password1', 'password2']
-
-
-class LoginForm(forms.Form):
-    username = forms.CharField(max_length=150)
-    password = forms.CharField(max_length=128)
