+++ backend/users/models.py
@@ -5,8 +5,7 @@ from django.db import models
 from PIL import Image
 
 from .constants import AVATAR_DEFAULT, AVATAR_MAX_HEIGHT, AVATAR_MAX_WIDTH
-from .utils import avatar_upload_to
-from .validators import validate_avatar_size
+from .utils import avatar_upload_to, validate_avatar_size
 
 
 class UserProfile(AbstractUser):
