+++ backend/users/utils.py
@@ -1,7 +1,9 @@
 import os
 import uuid
 
-from .constants import AVATAR_DEFAULT, AVATAR_UPLOAD_DIR
+from django.core.exceptions import ValidationError
+
+from .constants import AVATAR_DEFAULT, AVATAR_UPLOAD_DIR, AVATAR_MAX_SIZE_MB
 
 
 def avatar_upload_to(instance, filename):
@@ -10,3 +12,8 @@ def avatar_upload_to(instance, filename):
         new_filename = f"{uuid.uuid4()}.{ext}"
         return os.path.join(AVATAR_UPLOAD_DIR, new_filename)
     return AVATAR_DEFAULT
+
+
+def validate_avatar_size(file):
+    if file.size > AVATAR_MAX_SIZE_MB * 1024 * 1024:
+        raise ValidationError(f"File size should not exceed {AVATAR_MAX_SIZE_MB} MB.")
